Implement q_new and q_free
q_new creates a new queue and returns a pointer to its head.
It allocates memory for a struct list_head and initializes it.

q_free releases the memory occupied by the queue referenced by head.
It iteratively frees both the data in each element and the elements
themselves. Finally, it reconnects the nodes before and after
head and frees the head.

Change-Id: I29477f34472db4d05bfd33a74073fac6aade4f29