Implement q_insert_head and q_insert_tail
The difference between these two functions is the position of insertion.
First, check if head is a NULL pointer.
Next, allocate memory for the list entry @element and its data
(@element->data) to store the string. Then, copy the input string
into the entry.
Finally, insert the new entry in the appropriate position.

Change-Id: I8de0f1431c46a6bf32094ec7196117b3c9615998