Implement q_remove_head and q_remove_tail
The difference between these two functions is the position of removal.
If @head is NULL, return immediately.
If string pointer is not NULL, copy the data from the entry within
the buffer size. Otherwise, skip the copy operation.
Finally, remove the entry, reconnect the queue, initialize the entry,
and return its reference.

Change-Id: I9d56531b22f168ed1949737cebc4e364073ff584